remove unnecessary code and refactored some interal functions and moved some funcion to helpers file
