+++ labtool_ex2/src/labtool_ex2/helpers.py
@@ -1,0 +1,10 @@
+import math
+
+
+def round_up(n, decimals=0):
+    multiplier = 10 ** decimals
+    return math.ceil(n * multiplier) / multiplier
+
+
+def orderOfMagnitude(number):
+    return math.floor(math.log(number, 10))
